dev-amdgpu,stdlib: Make GPU VRAM size flexible
The size of GPU memory (VRAM) is current fixed to 16GiB due to a
hardcoded value for the memory size register. This can be a lot of
memory to simulate. In particular, for multiple GPU systems were
measuring communication overhead is the goal, we don't necessarily want
16GiB of memory for *every* GPU.

This changes the VRAM size from being hard-coded to 16GiB to use the
size set by a stdlib AbstractMemorySystem. The value of BAR0 (the PCIe
BAR aliasing VRAM memory) is set to match the size of the stdlib
AbstractMemorySystem. The sizes matching turns on a feature called
"large BAR" which allows direct access to VRAM using a PCIe address. The
AbstractMemorySystem size is also set to the memory size register which
is what officially tells the amdgpu Linux driver how much memory is
available.