stdlib: Update GPU driver command for mGPU and ROCm 6.2
Starting in ROCm 6.2 the driver cannot be simply modprobe'd due to lack
of ACPI support in gem5. Instead, a script on ROCm 6.2+ disk images is
used to handle all of the details of that ROCm version. The load command
checks for the new script first, then tries to modprobe for backward
compatibility, and exits otherwise.

For multiple GPU support, add an `lspci -v` command to confirm that
multiple PCIe GPUs can be seen.